+++ app/src/test/java/ua/com/radiokot/lnaddr2invoice/integration/logic/GetBolt11InvoiceUseCaseTest.kt
@@ -6,7 +6,7 @@ import org.koin.test.KoinTest
 import org.koin.test.get
 import ua.com.radiokot.lnaddr2invoice.di.commonModules
 import ua.com.radiokot.lnaddr2invoice.logic.GetBolt11InvoiceUseCase
-import java.math.BigDecimal
+import java.math.BigInteger
 
 class GetBolt11InvoiceUseCaseTest : KoinTest {
     @Test
@@ -18,7 +18,7 @@ class GetBolt11InvoiceUseCaseTest : KoinTest {
         }
 
         val invoice = GetBolt11InvoiceUseCase(
-            amountSat = BigDecimal(5),
+            amountSat = BigInteger.valueOf(5),
             callbackUrl = "https://livingroomofsatoshi.com/api/v1/lnurl/payreq/2adffa0b-67c1-4b50-a5be-4994b488efb1",
             jsonObjectMapper = get(),
             httpClient = get(),
